+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - SMTP envelope MAIL FROM address #16
 - SMTP exposes timeout #20 , attempts, delay
 - SMTP re-connect on every error #21
+- smtp.quit() exposed #5
 
 ## 1.5.3 (2021-08-02)
 - group syntax is no more reported as erroneous (ex: undisclosed-recipients:;)